Customise indentation for JSON-like databases
